CS: move mock classes to `test/doubles` directory [1]
Includes:
* Putting them in a logical path.
* Renaming the namespace.
* Adjusting any and all files in which the Mock class is being used.
